pyproject.toml

[tool.poetry]
name = "forum"
version = "0.1.0"
description = ""
authors = ["LeeJun"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
pytest-django = "^4.4.0"
black = "^21.9b0"
autopep8 = "^1.5.7"
flake8 = "^3.9.2"
django-stubs = "^1.9.0"
toolz = "^0.11.1"
pytest-mock = "^3.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
